+++ modules/Yazdan/Game/src/App/Http/Controllers/Game/GameController.php
@@ -17,7 +17,7 @@ class GameController extends Controller
     {
         $this->authorize('manage', Game::class);
         $games = Game::latest()->paginate(10);
-        return view('Game::admin.index',compact('games'));
+        return view('Game::admin.index', compact('games'));
     }
 
     public function create()
@@ -30,10 +30,11 @@ class GameController extends Controller
     {
         $this->authorize('manage', Game::class);
         if (isset($request->media)) {
-              $images = MediaFileService::publicUpload($request->media);if($images == false){
-            newFeedbacks('نا موفق','فرمت فایل نامعتبر میباشد','error');
-            return back();
-        }
+            $images = MediaFileService::publicUpload($request->media);
+            if ($images == false) {
+                newFeedbacks('نا موفق', 'فرمت فایل نامعتبر میباشد', 'error');
+                return back();
+            }
             $request->request->add(['media_id' => $images->id]);
         }
         GameRepository::store($request);
@@ -56,9 +57,9 @@ class GameController extends Controller
     public function edit($id)
     {
         $this->authorize('manage', Game::class);
+        $statuses = GameRepository::$statuses;
         $game = GameRepository::findById($id);
-
-        return view('Game::admin.edit', compact('game'));
+        return view('Game::admin.edit', compact('game','statuses'));
     }
 
     public function update($id, GameRequest $request)
@@ -67,17 +68,17 @@ class GameController extends Controller
         $game = GameRepository::findById($id);
 
         if ($request->hasFile('media')) {
-
+            $images = MediaFileService::publicUpload($request->media);
+            if ($images == false) {
+                newFeedbacks('نا موفق', 'فرمت فایل نامعتبر میباشد', 'error');
+                return back();
+            }
             if ($game->media) {
                 $game->media->delete();
             }
-              $images = MediaFileService::publicUpload($request->media);if($images == false){
-            newFeedbacks('نا موفق','فرمت فایل نامعتبر میباشد','error');
-            return back();
-        }
             $request->request->add(['media_id' => $images->id]);
         } else {
-            if ($game->media) {
+            if ($game->media && $game->media->id) {
                 $request->request->add(['media_id' => $game->media->id]);
             }
         }
@@ -91,8 +92,7 @@ class GameController extends Controller
         $this->authorize('manage', Game::class);
         $game = GameRepository::findById($gameId);
         $levels = LevelRepository::GameLevelPaginate($gameId, 15);
-        return view('Game::admin.details',compact('game','levels'));
-
+        return view('Game::admin.details', compact('game', 'levels'));
     }
 
 
@@ -101,6 +101,6 @@ class GameController extends Controller
         $this->authorize('manage', Game::class);
         $game = RecordRepository::GameRecordsPaginate($gameId);
         $records = $game->records()->latest()->paginate(20);
-        return view('Game::admin.records',compact('game','records'));
+        return view('Game::admin.records', compact('game', 'records'));
     }
 }
